Extract Post methods to a PostAnalyzer class
Extract specs as well
