+++ .gitignore
@@ -35,6 +35,7 @@ config/discourse.pill
 # Ignore all logfiles and tempfiles.
 /log/*.log
 /tmp
+/logfile
 
 # Ignore Eclipse .project file
 /.project
