+++ app/models/post.rb
@@ -4,7 +4,7 @@ require_dependency 'rate_limiter'
 require_dependency 'post_revisor'
 require_dependency 'enum'
 require_dependency 'trashable'
-require_dependency 'post_analyser'
+require_dependency 'post_analyzer'
 
 require 'archetype'
 require 'digest/sha1'
@@ -12,7 +12,6 @@ require 'digest/sha1'
 class Post < ActiveRecord::Base
   include RateLimiter::OnCreateRecord
   include Trashable
-  include PostAnalyser
 
   versioned if: :raw_changed?
 
@@ -90,10 +89,10 @@ class Post < ActiveRecord::Base
     Digest::SHA1.hexdigest(raw.gsub(/\s+/, "").downcase)
   end
 
-  def cooked_document
-    self.cooked ||= cook(raw, topic_id: topic_id)
-    @cooked_document ||= Nokogiri::HTML.fragment(cooked)
-  end
+  #def cooked_document
+    #self.cooked ||= cook(raw, topic_id: topic_id)
+    #@cooked_document ||= Nokogiri::HTML.fragment(cooked)
+  #end
 
   def reset_cooked
     @cooked_document = nil
@@ -104,18 +103,31 @@ class Post < ActiveRecord::Base
     @white_listed_image_classes ||= ['avatar', 'favicon', 'thumbnail']
   end
 
-  # How many images are present in the post
-  def image_count
-    return 0 unless raw.present?
-
-    cooked_document.search("img").reject do |t|
-      dom_class = t["class"]
-      if dom_class
-        (Post.white_listed_image_classes & dom_class.split(" ")).count > 0
-      end
-    end.count
+  def post_analyzer
+    @post_analyzer ||= PostAnalyzer.new(self)
   end
 
+  %w{raw_mentions linked_hosts image_count link_count raw_links}.each do |attr|
+    define_method(attr) do
+      post_analyzer
+      @post_analyzer.send(attr)
+    end
+  end
+
+  # What we use to cook posts
+  def cook(*args)
+    cooked = PrettyText.cook(*args)
+
+    # If we have any of the oneboxes in the cache, throw them in right away, don't
+    # wait for the post processor.
+    dirty = false
+    result = Oneboxer.apply(cooked) do |url, elem|
+      Oneboxer.render_from_cache(url)
+    end
+
+    cooked = result.to_html if result.changed?
+    cooked
+  end
 
   # Sometimes the post is being edited by someone else, for example, a mod.
   # If that's the case, they should not be bound by the original poster's
@@ -236,20 +248,6 @@ class Post < ActiveRecord::Base
     Post.excerpt(cooked, maxlength, options)
   end
 
-  # What we use to cook posts
-  def cook(*args)
-    cooked = PrettyText.cook(*args)
-
-    # If we have any of the oneboxes in the cache, throw them in right away, don't
-    # wait for the post processor.
-    dirty = false
-    result = Oneboxer.apply(cooked) do |url, elem|
-      Oneboxer.render_from_cache(url)
-    end
-
-    cooked = result.to_html if result.changed?
-    cooked
-  end
 
   # A list of versions including the initial version
   def all_versions
